Moved ContestPermissionAdminForm to forms, added autocomplete searchfield for adding permissions for teachers to a contest
