+++ oioioi/contests/views.py
@@ -695,3 +695,7 @@ def reattach_problem_confirm_view(request, problem_instance_id, contest_id):
         'contests/reattach_problem_confirm.html',
         {'problem_instance': problem_instance, 'destination_contest': contest},
     )
+
+def get_teacher_names(request):
+    queryset = User.objects.filter(teacher__isnull=False)
+    return get_user_hints_view(request, 'substr', queryset)
